Update ScanKeys.txt
Keys are ommitted to keep account access private. but if you want test it out and use your own, just put the keys to another twitter account in, with a space between the '=" and the keys. this is the keys for the account that will be used only to access the api for tweet scraping purposes, for this purpose I used my personal twitter account, but its really irrelevant what account goes here as the scrape bot does not post