Update realDonaldTrumpBotKeys.txt
keys are ommitted to keep account access private, you can have the bot access any number of accounts to post and comment. Just put the account name of the figure you are parodying, followed by 'BotKeys' as a filename, create an account to be your bot, input the keys from the account a file formatted like this, run the scraper, and the code will generate the bot for you(as many as you want).